Refs #6578. Taking GuidPrefix_t to its own header.
